backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
supabase>=2.0.0
google-auth>=2.0.0
google-api-python-client>=2.0.0
PyJWT>=2.0.0
email-validator>=2.0.0
slowapi>=0.1.9
gunicorn>=21.0.0
